fix(trusty-mpm): split session_launch tests to stay under the 1500-SLOC test cap
session_launch/tests.rs grew to 1507 SLOC once the #3427 scaffolding-gitignore
wiring tests landed, tripping CI's ratcheted line-cap check. Move the two new
tests (prepare_session_gitignores_scaffolding_when_project_is_git_repo,
prepare_session_skips_gitignore_when_project_is_not_git_repo) to a new sibling
tests_scaffold_gitignore.rs, mirroring the existing tests_roster.rs /
native_mcp_tests.rs / custom_mcp_tests.rs split convention for this directory.
No behavior change; `scripts/check_line_cap.sh` now passes clean.

Refs #3427
